requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.4
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2025.5.1
gdown==5.2.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imgviz==1.7.6
Jinja2==3.1.6
kiwisolver==1.4.8
labelme==5.8.2
lazy_loader==0.4
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
natsort==8.4.0
networkx==3.5
numpy==2.2.6
onnxruntime==1.22.1
opencv-python==4.12.0.88
osam==0.2.4
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
sympy==1.14.0
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.165
ultralytics-thop==2.0.14
urllib3==2.5.0
